created file for experimental conditions;
